Close and dispose test database setup connections
Without explicitly closing and disposing the connections can linger in a
non-deterministic fashion and cause intermittent "too many connection"
errors in the tests.

These changes were diagnosed and then tested by lowering test
concurrency to 1 and lowering available postgres connections to 10
and forcing the problem early in the test run.

Change-Id: I8badf2a3094f2d27d381bd178c6d579f2a5c5ffe