Add basic gate functional testing jobs for aodh
This provides a way to run functional tests from the gate, initially
just a simple set of gabbi tests which exercise the API. Once this
has proven to be working we can add more.

The model here is a hybrid of what was started for ceilometer and
what was used in gnocchi with some adjustments to allow for some
local testing.

The tests produced by test_gabbi_live will only run if
AODH_SERVICE_URL is set to the alarming endpoint. A keystone auth
token is needed in AODH_SERVICE_TOKEN. post_test_hook.sh
makes this happen if we are in a devstack-gate environment.
Otherwise they must be set by the caller.

The expectation is that we'll stick more test in tests/functional as
we have time, energy and brains.

Note that the layout of the functional tests assumes that the tests
will skip (or not even generate tests) if a functional environment
is not in place, such as when the unit tests are run.

Change-Id: Ia40e9c6b956209ecd0086ca861a90ce0c92b7b10