Keep alarm other attrs constantly after evaluating an alarm
When evaluating a threshold alarm, the alarm rule of the *alarm* object
will be changed, because in the _statistics() method, the
alarm.rule['query'] is referenced and extended. When the alarm being
refreshed, the *alarm* object will be updated to db in _refresh()
method, that will cause the alarm's query definition being extended
repeatedly.

Change-Id: Iefb4646ee9d07a272550f13a91ae60220032e4db