gnocchi: percent_of_overlap=0 for agg. alarms
In case of a heat autoscaling stack decide to delete an instance,
the gnocchi metrics associated to this instance will be no more
updated and when the alarm will ask for the aggregation, gnocchi
will raise a 'No overlap' exception.
So temporary set 'percent_of_overlap' to 0 to disable the
gnocchi checks about missing points.

And adds some docs about the importance of filter out resource
with ended_at != 'null'.

For more detail see:
  https://bugs.launchpad.net/gnocchi/+bug/1479429

Related bug: #1479429

Change-Id: I528d21c048d84b9552febd08f787812c668983ab