Make ConnectionRetryTest more reliable
The previous test relied on mocks of retrying.time.sleep. This is
the same sleep as time.sleep so there's a chance, if eventlet is in
context, that something else can call that.

So instead the test now checks to see how many times should_reject
is called. This is the same as the max retries that will be
attempted.

Also the retry wait time has been set to something very small so the
test doesn't hang around sleeping uselessly.

Change-Id: Ib85fff55588f7183d4e27a8a4769f647aeaa0ce3
Closes-Bug: #1482722