event-alarm: delete debug message to show # of alarms
This patch delete the debug log mssage to show the number of alarms
found at the beginning of event evaluation. Developer still can see
those alarms in the log messages generated in each alarm evaluation.

The alarm storage api returns generator, so it should be safe that
we don't touch the returned value.

Change-Id: I6d61be5ab8ae2ee53a1528477a5686a5e7fa33d6
Closes-Bug: #1492098