utils: move code where it's actually used and remove
This patch removes the aodh.utils module by moving the code it contains
where it is actually used.

All the code it contains actually belong to some other places as it's
not shared anymore. This cleans up a bit the code by removing the
module, and makes sure we don't carry dead code in the future.

Change-Id: Ic31aa82d471fc46d64ad546c13f677d897d169d7