Move the content of ReleaseNotes to README.rst
For now, we have added reno for release notes management, and a folder
"releasenotes" was added in aodh. Because there is a file named
"ReleaseNotes" to provide release note location and aodh home, it will
broken some git operations in case insensitive OS (e.g. Windows).
This change simply move the content of ReleaseNotes to README.rst to avoid
the problem.

Change-Id: I30ea4a5d8cb4450ca76ec5e834ccf07cd996de77