[Trivial] Remove AODH_API_LOG_DIR option for devstack
This option is used for aodh-api when AODH_DEPLOY=werkzeug,
and create a dedicate directory for aodh-api log.

However, there is no such need, because for such case, aodh-api
is just a normal process like aodh-evaluator. We should let
oslo.log decide where to store the log, image that user configures
a path and we override in installation process, then he needs to hack
into screenrc to adjust the path, it is not friendly.

Here I propose to put the api log files in same directory as
aodh-evaluator, by default, it will be /opt/stack/logs.

Change-Id: If758cb7ddcc0087ed495721eeccf4723900e2ef3