Enable aodh service to be multi-processes
aodh-notifier is stateless, it is naturally can be multi-processes.
aodh-evaluator can be multi-processes because it by default runs
in workload_partition mode.
aodh-listener shares nothing between workers, an event can only
be consumed by a single worker, and sql provides transaction ability
hence it is safe to use multi-processes as well.

Change-Id: I5a8051909be8626cd9d037992132740d03110892