log alarm rest notifier response
Currently we do nothing after post the rest action, so we don't
know the actual result of it, except it raises exception. This
is not right because there is no exception even for 400 and 500
error code.

This patch adds a log message at info level to enable cloud
administrator track and audit the alarm notify result.

like:

  Notifying alarm <537944be-1dfd-48fe-b272-2ae0b04e04e1>
  gets response: 500 Internal Server Error.

Change-Id: I56ab6256260697d2e9fe66dfc30e0adbb4024814