Add pre-signed support for zaqar notifier
Currently the zaqar notifer driver will create a queue per alarm
in service tenant, as a result, end user can't access the queue
from zaqar. Which makes it's hard to add a subscriber like mistral
because it needs a complex subscription options. The patch will
provide a capabability for end user let aodh forward the alarms
to user's queue instead of the queue created automatically in
service tenant. And this patch could benefit any 3rd party which
interested in the alarm information.

Change-Id: I4b843b39c8a13bd40aa2923a62ba681c81e06e9a