Restore trust workaround for legacy auth
Before mitaka, we were not supporting keystone v3 and we were
expecting to have the auth_url ending with /v2.0 to ensure
v2 keystone API. We were using a workaround to force v3 for trust
keystone client, and when we moved to keystoneauth1 we removed this workaround
but it's still needed.

This change restores it. It ensures the endpoint is a v3 endpoint when
the user enter a v2 endpoint and set domain to 'default' that was the
only that worked before, because we were not really support keystone v3.

This is not a backport because it fixes a bug in deprecated (and already removed from master) code.

Closes-bug: #1620499
Change-Id: Idf474348ab4d49073dc5561caf90e151c4138819