Change tooz and hbase requirement
This change are merged together to fix the gating

hbase part:
The lib API have changed since 1.0 release, Aodh uses API
that have been removed/renamed.

So we cap requirement to <1.0 to ensure Aodh does not
break.

tooz part:
During the backport of https://review.openstack.org/#/c/326573/, we bump
to 1.28. This version is too high for liberty.

https://review.openstack.org/#/c/310337/ explains that only 1.19.0 is
required, that's perfect this is the version required by liberty.

Change-Id: I9b50ea959d8765ed0a53e0994f60774284a982cd