Add http_proxy_to_wsgi to api-paste
This sets up the HTTPProxyToWSGI middleware in front of Aodh. The
purpose of thise middleware is to set up the request URL correctly in
case there is a proxy (For instance, a loadbalancer such as HAProxy)
in front of Aodh.

So, for instance, when TLS connections are being terminated in the
proxy, and one tries to get the versions from the / resource of
Aodh, one will notice that the protocol is incorrect; It will show
'http' instead of 'https'. So this middleware handles such cases.
Thus helping Keystone discovery work correctly.

The HTTPProxyToWSGI is off by default and needs to be enabled via a
configuration value.

Change-Id: If2ada8a94c8e1ceacd4509605b4cd766a78f71d5
Closes-Bug: #1590608