modernise gabbi usage
as our borderline millenial hero progressed through the land of dried
bread, he found himself at the gates of hell. the flames of unreadable
wsme code overwhelmed him. as a lazy protagonist, he decided to ignore
wsme and just deal with cleaning up gabbi. this work came a bit easier
as it was quite similar to his work in panko and thus, he took gabbi
v1.7, and drop some of the redundancy.

Change-Id: I19dfb39ea8b9a4fc4063e32ce19b62d8ce74f939