Gnocchi: don't fail to create valid alarm
When the Gnocchi resource or metric doesn't exists yet, we refuse
to create the alarm and return a confusing 404.

This change returns 200, in such case and removes the need to have the
Gnocchi resource and metric existing, we just trust that the id/name
provided by the use will exists.

Depends-On: I810897e750c4b5a42bb3ada2f8a2488dbee609f4
Change-Id: I404daf0c30d4b3ced239a94b98d8a62bad71a6b6