Remove deprecated oslo.config messages.
DeprecationWarning: Using the 'enforce_type' argument is deprecated in
version '4.0' and will be removed in version '5.0': The argument
enforce_type has changed its default value to True and then will be
removed completely.

This option have never been useful, and now the default is already True.
So we don't need to raise requirement for this.

Change-Id: I042a5d98748e2b2c6ee8307340aa6056bf0a1dfd