Fix aodh-config-generator
This program does not work:
  Unable to build sample configuration file: Failed to find some config files: /home/jenkins/workspace/gate-aodh-python27-ubuntu-xenial/aodh/cmd/aodh-config-generator.conf

But the gate did not catch it because the returned value is still 0. This
patches fixes that.

It also fix the location of the aodh-config-generation.conf file to be in the
proper place, and remove the default output location since there is no such
directory in $PWD by default, obviously

Change-Id: Ib63bec174de0c3108d42ad0623baba96c6cbe78b
Closes-Bug: #1702476