Implement policy in code (1)
This commit will prepare for implementing policies in code[1]
that mean with oslo.policy, we can define all of default rules
in code base and only update some rules via policy file if need.

To do that, we can move "rule by rule" into code base.
In this change, we will:

- Define some common rules in code base.
- Register them into policy engine to use as default policy rules
- Remove them out of policy.json file, so if operators want to
update rules, they should define them in policy file like the way
we done with config option with oslo.config.

Summary, with any rules that we defined in codebase, we no need to keep
them in policy file anymore if we don't want to customize them.
And everything still work well meanwhile we didn't define rules
in policy file with falling back to default rules.

[1] https://governance.openstack.org/tc/goals/queens/policy-in-code.html

Change-Id: Ide8f581b9adb6701eeb9b1f5293307dcee3dd9c3
Co-authored-By: Hieu LE <hieulq@vn.fujitsu.com>