Fix releasenotes builds
The updated releasenote and requirements jobs expect requirements for
docs in one of the requirements files. Move requirements from setup.cfg
to doc/requirements.txt file.

Update also releasenotes build to use same command as in jobs - build
with warnings as errors.

See also
https://governance.openstack.org/tc/reference/project-testing-interface.html#documentation

Fix failure of building with warnings as errors:
Replace the instructions for querying for release notes for newton
with the results of a reno report because some of the broken release notes
are on branches that have been closed and we have no way to edit them.

Depends-On: I994aa3cf173ec8a9412239544e3269aca17adc25
Change-Id: If8df8a6e8cad2d9247f07c9ed7a71f29b19f7d51