Implement secure RBAC for alarms and quota policies
This commit updates the default policies to include supports for
enhanced token scope checking and default roles, including a read-only
role for GET APIs.

This is part of a broader change across OpenStack projects to
consolidate common personas and provide a more secure and consistent
authorization experience for end users and operators.

Change-Id: If17dbfec10302c9c57d814cb406faee1a6efc59a