Load api-paste.ini from configuration directories first
The api-paste.ini file defines WSGI definition of API. This file is
regarded as one configuration file and sometimes users need to modify
the file, for example to enable an additional middleware.

However currently aodh-api by default expects the file is located in
the aodh source directory and users should set the [api] paste_config
parameter in addition to copy and modify the file.

This change makes the aodh-api process search the api-paste.ini file
from configuration directories automatically, so that users can more
easily customize the file by putting the modified file in /etc/aodh
(or any other config directories).

If the paste_config parameter is defined as a full path then aodh-api
directly loads the file from the specified path. If the file is not
found then the default file in the aodh source directory is used.

Change-Id: I4a6194fbf6a5317a523d75e3fb6bd2fd055f5100