Ignore Gnocchi API error when the metric is not yet created
Change 74eadfbd58359b7ebe9e1e40ae6b6ff245146bb8 replaced usage of
metric aggregate API by Dynamic aggregate API.

However the Dynamic aggregate API returns a different response code
(not 404, but 400) in Gnocchi <= 4.4 and this is causing the direct API
error when the metric associated with the aggregation does not exist.

This change adds 400 to the ignored response codes so that Aodh api
does not return error even when the associated metric is not yet
created in Gnocchi.

Closes-Bug: #1974682
Change-Id: Ief75ccdc0fe5c6a197fc05bc481a5be096447558