Fix compatibility with oslo.db 12.1.0
oslo.db 12.1.0 has changed the default value for the 'autocommit'
parameter of 'LegacyEngineFacade' from 'True' to 'False'. This is a
necessary step to ensure compatibility with SQLAlchemy 2.0. However,
Aodh is currently relying on the autocommit behavior and needs changes
to explicitly manage sessions. Until that happens, we need to override
the default.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I129b54b06e1a5c3fafd2a03f295cc1cfd9c9e33c