Add evaluation results counters to /metrics
This change adds code for collection and storage of
evaluation result counters. The values of the counters can
then be retreived through a new /metrics api endpoint.

Counters are collected and served for each alarm separately. Only
metrics for the current project will be shown on the /metrics endpoint.
To enable this feature set "enable_evaluation_results_metrics"
to true inside [DEFAULT].

Change-Id: I3fceb9a244f248f1e62a1267408c82d1b35a4d37