Revert "Disable GLOBAL_VENV"
This reverts commit 8de7585f374097bcb4910bfbed34b6c721490db9.

Reason for revert:
Current devstack disables GLOBAL_VENV automatically in CentOS or Rocky
so we don't have to override the option. GLOBAL_VENV=True is required
to migrate the jobs to Ubuntu 24.04 (Noble) .

This also changes the default API deployment method to uwsgi because
uwsgi is now globally used in CI and only works for global venv
method.

Conflicts:
	.zuul.yaml

Change-Id: Iec6ab6b4fcdb60caa67197acb5ada0678c0309ce