Add wsgi module
The wsgi_script feature is being removed because of some changes in
underlying python packaging tooling. This makes aodh to vendor the wsgi
module which can be used instead of the wsgi script, according to
the proposed community goal[1]. The existing wsgi scripts are kept now
for smooth transition.

The exsiting app.wsgi file is deprecated in favor of the new wsgi
module to follow the standard pattern to vendor the wsgi module.

[1] https://review.opendev.org/c/openstack/governance/+/902807

Change-Id: I7ce471c7fbce960ba92bfd18203a73454088314e