Deprecate the prometheus_disable_rbac cfg option
The current implementation of the observabilityclient rbac feature
can't be used by services, it's usable by regular users only.
Using this feature can cause some Prometheus type alarms to not
work correctly and end up in the "insufficient data" state.

This change deprecates this config option and hardcodes it to
"True" for disabling the feature.

Closes-Bug: #2106029
Change-Id: I2146e8e753fd7b1214ff583d9d85bbd71bd36fed