rely on devstack to install uwsgi
installation of uwsgi from pypi is flaky
as the compilation of the uwsgi c module sometimes
fails for non determinisic reasons.

Devstack already will install uwsgi for us and
it will prefer to use the distro package if possible avoiding
the need to compile it in the job. this is both faster and
more reliable reducing rechecks.

Closes-Bug: #2126581
Change-Id: I6f322df9a72316f97f8e67427d9cbdc4e2c01505
Signed-off-by: Sean Mooney <work@seanmooney.info>